board 작성한 글이 있는 멤버 삭제, 본인의 정보만 수정

+++ src/main/java/com/proj1/controller/memberController.java
@@ -3,6 +3,7 @@ package com.proj1.controller;
 import com.proj1.domain.Member;
 import com.proj1.service.MemberService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -43,8 +44,11 @@ public class memberController {
     }
 
     @PostMapping("remove")
-    public String remove(Integer id) {
-        service.remove(id);
+    public String remove(Integer id, Authentication authentication) {
+        if (service.hasAccess(id, authentication)) {
+            service.remove(id);
+        }
+
 
         return "redirect:/member/signup";
     }
